Add files via upload
I'm working on this project to ensure the office runs smoothly and to prevent health issues. I need to figure out who enjoys working in the office, who prefers working from home, and when people usually take sick days. The reason for having a hybrid job setup is to save money on renting a big building and to create fun events for employees to maintain a positive company culture. By understanding these preferences, the company can plan events at the office when most people are likely to be there and take precautions if many employees get sick. This way, the can keep things running smoothly and make sure everyone stays healthy and happy.